refactor: 로또 생산 책임 온전히 LottoFactory에 옮김
