+++ src/test/java/controller/LottoPurchaseControllerTest.java
@@ -44,7 +44,7 @@ class LottoPurchaseControllerTest {
         Assertions.assertTrue(inputView.getUserCashInputCalled);
         Assertions.assertEquals(2, lottoFactory.generateLottoCalledCount);
         Assertions.assertEquals(2, controller.wrapLottoIntoDtoCallCount);
-        Assertions.assertTrue(controller.checkPriceHigherThanSingleLottoPriceCalled);
+        Assertions.assertTrue(lottoFactory.checkPriceHigherThanSingleLottoPriceCalled);
 
     }
 }
