+++ src/test/java/controller/mock/MockLottoPurchaseController.java
@@ -9,9 +9,7 @@ import view.InputView;
 import view.OutputView;
 
 public class MockLottoPurchaseController extends LottoPurchaseController {
-    public int getLottoCallCount = 0;
     public int wrapLottoIntoDtoCallCount = 0;
-    public boolean checkPriceHigherThanSingleLottoPriceCalled= false;
 
     public MockLottoPurchaseController(
             LottoBatch lottoBatch,
@@ -22,20 +20,8 @@ public class MockLottoPurchaseController extends LottoPurchaseController {
     }
 
     @Override
-    protected void generateLottoByPrice(int userCashInput) {
-        getLottoCallCount++;
-        super.generateLottoByPrice(userCashInput);
-    }
-
-    @Override
     protected LottoDto wrapLottoIntoDto(Lotto lotto) {
         wrapLottoIntoDtoCallCount++;
         return super.wrapLottoIntoDto(lotto);
-    }
-
-    @Override
-    protected void checkPriceHigherThanSingleLottoPrice(int price) {
-        checkPriceHigherThanSingleLottoPriceCalled = true;
-        super.checkPriceHigherThanSingleLottoPrice(price);
     }
 }
